Create LinearFeedbackShiftRegister.java
Linear Feedback Shift Register A shift register is an array of bits that supports
a shift operation which moves every bit one slot to the left. For example, below is an example of
a shift register initially containing the seed \10101111". It is shifted twice. Each time it is shifted,
the leftmost bit is dropped, every other bit is moved one place to the left, and the rightmost bit is
replaced with a `0'. Notice that after it is shifted 8 times, the shift register will be all zeros.
A linear feedback shift register is a special type of shift register that, instead of setting the rightmost
bit to zero, updates the rightmost bit with some linear function of the other bits. That is, when a
shift operation occurs, it feeds back some information into the low-order bit. We will build a linear
feedback shift register based on the exclusive-or (XOR) function. (In Java, (a^b) calculates the
XOR of a and b.)
Our linear feedback shift register takes two parameters: a size and a tap. The size refers to the
number of bits. The tap refers to one of the bits and must be between 0 and size􀀀1 (inclusive).
Every time a shift operation occurs, the following four things happen: (1) The feedback bit is
calculated as the XOR of the leftmost bit and the tap bit. (2) The leftmost bit is dropped. (3)
Every bit is moved one slot to the left. (4) The rightmost bit is set to the feedback bit. Here are
two examples of a linear feedback shift register in action. In both cases, the size is 8 and the tap
is 4.

Implement a linear feedback shift register that implements the ILFShiftRegister
interface:
public interface ILFShiftRegister {
public void setSeed(int[] seed);
public int shift();
public int generate(int k);
}
The interface requires that the following methods be supported:
 void setSeed(int[] seed): sets the shift register to the specied initial seed. The initial
seed is specied as an array of 0's and 1's (represented as integers). If the seed contains any
other value, that is an error. (Recall that the seed is the initial set of bits stored in the shift
register.)
 int shift(): executes one shift step and returns the rightmost bit of the resulting register.
 int generate(int k): extracts k bits from the shift register. It executes the shift operation
k times, saving the k bits returned. It then converts these k bits from binary into an integer.
For example: if generate is called with a value of 3, then it calls shift 3 times. If the shift
operations return 1 and then 1 and then 0, then generate returns the value 6 (i.e., \110" in
binary).
Your implementation should also support the following constructor:
ShiftRegister(int size, int tap)
This constructor initializes the shift register with its size and with the appropriate tap. Submit
your implementation.